+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to `cybercog/laravel-paket` will be documented in this file.
 
 - ([#39]) Add jobs deletion
 
+### Fixed
+
+- ([#42]) Lock UI when Job in progress
+
 ## [1.3.0]
 
 ### Added
@@ -47,6 +51,7 @@ All notable changes to `cybercog/laravel-paket` will be documented in this file.
 [1.2.0]: https://github.com/cybercog/laravel-paket/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/cybercog/laravel-paket/compare/1.0.0...1.1.0
 
+[#42]: https://github.com/cybercog/laravel-paket/pull/42
 [#35]: https://github.com/cybercog/laravel-paket/pull/35
 [#33]: https://github.com/cybercog/laravel-paket/pull/33
 [#32]: https://github.com/cybercog/laravel-paket/pull/32