+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to `cybercog/laravel-paket` will be documented in this file.
 - ([#48]) Added repositories tab
 - ([#51]) Added refetch data button in top menu
 - ([#52]) Added polling to get realtime job output
+- ([#53]) Added requirement suggestions from Packagist.org
 
 ## [1.4.0]
 
@@ -59,6 +60,7 @@ All notable changes to `cybercog/laravel-paket` will be documented in this file.
 [1.2.0]: https://github.com/cybercog/laravel-paket/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/cybercog/laravel-paket/compare/1.0.0...1.1.0
 
+[#53]: https://github.com/cybercog/laravel-paket/pull/53
 [#52]: https://github.com/cybercog/laravel-paket/pull/52
 [#51]: https://github.com/cybercog/laravel-paket/pull/51
 [#48]: https://github.com/cybercog/laravel-paket/pull/48