+++ CHANGELOG.md
@@ -64,6 +64,7 @@ All notable changes to `cybercog/laravel-paket` will be documented in this file.
 [1.2.0]: https://github.com/cybercog/laravel-paket/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/cybercog/laravel-paket/compare/1.0.0...1.1.0
 
+[#55]: https://github.com/cybercog/laravel-paket/pull/55
 [#53]: https://github.com/cybercog/laravel-paket/pull/53
 [#52]: https://github.com/cybercog/laravel-paket/pull/52
 [#51]: https://github.com/cybercog/laravel-paket/pull/51